pyproject.toml
[project]
name = "car-detector"
version = "0.1.0"
description = ""
authors = [
    {name = "Fernando Daza",email = "fernando.daza.18@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain[google-genai] (>=1.0.3,<2.0.0)",
    "fastapi[standard] (>=0.120.2,<0.121.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "ultralytics (>=8.3.224,<9.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
